refactor(scripts): allow importing without extension
