+++ scripts/src/files/file-reader.ts
