+++ scripts/src/files/file-writer.ts
