--- scripts/src/files/json-file-type.mts
@@ -1,4 +1,0 @@
-import { FileType } from './file-type.mts'
-import { JsonFile } from './json-file.mts'
-
-export const JSON_FILE_TYPE = new FileType('.json', JsonFile, JsonFile)
