+++ scripts/src/generators/content-generators.ts
@@ -1,10 +1,10 @@
-import { ResourceCollectionListGenerator } from '../resources/resource-collection-list-generator.mts'
-import { ProjectListItemExtraDataGenerator } from './project-list-item-extra-data-generator.mts'
-import { Resource } from '../resources/resource.mts'
-import { isMain } from '../utils/is-main.mts'
-import { Log } from '../utils/log.mts'
-import { Collections } from './collections.mts'
-import { IMAGES_FILE_BASENAME } from '../../../src/app/common/files.ts'
+import { ResourceCollectionListGenerator } from '../resources/resource-collection-list-generator'
+import { ProjectListItemExtraDataGenerator } from './project-list-item-extra-data-generator'
+import { Resource } from '../resources/resource'
+import { isMain } from '../utils/is-main'
+import { Log } from '../utils/log'
+import { Collections } from './collections'
+import { IMAGES_FILE_BASENAME } from '../../../src/app/common/files'
 
 export class ContentGenerators {
   public readonly collections: Collections
