+++ scripts/src/utils/get-repository-root-dir.ts
