+++ scripts/src/utils/is-main.ts
