+++ scripts/src/utils/log.ts
